实现Ota设备注册逻辑
