+++ internal/app/server/websocket/websocket_server.go
@@ -2,8 +2,6 @@ package websocket
 
 import (
 	"context"
-	"encoding/base64"
-	"encoding/json"
 	"fmt"
 	"io"
 	"net/http"
@@ -20,7 +18,6 @@ import (
 	userconfig "xiaozhi-esp32-server-golang/internal/domain/config"
 	utypes "xiaozhi-esp32-server-golang/internal/domain/config/types"
 	"xiaozhi-esp32-server-golang/internal/domain/mcp"
-	"xiaozhi-esp32-server-golang/internal/util"
 	log "xiaozhi-esp32-server-golang/logger"
 )
 
@@ -68,6 +65,7 @@ func (s *WebSocketServer) Start() error {
 	// 注册路由处理器
 	http.HandleFunc("/xiaozhi/v1/", s.handleWebSocket)
 	http.HandleFunc("/xiaozhi/ota/", s.handleOta)
+	http.HandleFunc("/xiaozhi/ota/activate", s.handleOtaActivate)
 	http.HandleFunc("/xiaozhi/mcp/", s.handleMCPWebSocket)
 	http.HandleFunc("/xiaozhi/api/mcp/tools/", s.handleMCPAPI)
 	http.HandleFunc("/xiaozhi/api/vision", s.handleVisionAPI) //图片识别API
@@ -304,102 +302,6 @@ func (s *WebSocketServer) handleWebSocket(w http.ResponseWriter, r *http.Request
 				break
 			}
 		}
-	}
-}
-
-func (s *WebSocketServer) handleOta(w http.ResponseWriter, r *http.Request) {
-	//获取客户端ip
-	ip := r.Header.Get("X-Real-IP")
-	if ip == "" {
-		ip = r.Header.Get("X-Forwarded-For")
-	}
-	if ip == "" {
-		ip = r.RemoteAddr
-	}
-
-	//从header头部获取Device-Id和Client-Id
-	deviceId := r.Header.Get("Device-Id")
-	clientId := r.Header.Get("Client-Id")
-
-	if deviceId == "" || clientId == "" {
-		log.Errorf("缺少Device-Id或Client-Id")
-		http.Error(w, "缺少Device-Id或Client-Id", http.StatusBadRequest)
-		return
-	}
-
-	deviceId = strings.ReplaceAll(deviceId, ":", "_")
-
-	//根据ip选择不同的配置
-	clientIp := r.Header.Get("X-Real-IP")
-	if clientIp == "" {
-		clientIp = r.Header.Get("X-Forwarded-For")
-	}
-	if clientIp == "" {
-		clientIp = r.RemoteAddr
-	}
-
-	otaConfigPrefix := "ota.external."
-	//如果ip是192.168开头的，则选择test配置
-	if strings.HasPrefix(clientIp, "192.168") || strings.HasPrefix(clientIp, "127.0.0.1") {
-		otaConfigPrefix = "ota.test."
-	} else {
-		otaConfigPrefix = "ota.external."
-	}
-
-	mqttInfo := getMqttInfo(deviceId, clientId, otaConfigPrefix, ip)
-	//密码
-	respData := &OtaResponse{
-		Websocket: WebsocketInfo{
-			Url:   viper.GetString(otaConfigPrefix + "websocket.url"),
-			Token: viper.GetString(otaConfigPrefix + "websocket.token"),
-		},
-		Mqtt: mqttInfo,
-		ServerTime: ServerTimeInfo{
-			Timestamp:      time.Now().UnixMilli(),
-			TimezoneOffset: 480,
-		},
-
-		Firmware: FirmwareInfo{
-			Version: "0.9.9",
-			Url:     "",
-		},
-	}
-
-	w.Header().Set("Content-Type", "application/json")
-	if err := json.NewEncoder(w).Encode(respData); err != nil {
-		log.Errorf("OTA响应序列化失败: %v", err)
-		http.Error(w, "内部服务器错误", http.StatusInternalServerError)
-		return
-	}
-	return
-}
-
-func getMqttInfo(deviceId, clientId, otaConfigPrefix, ip string) *MqttInfo {
-	if !viper.GetBool(otaConfigPrefix + "mqtt.enable") {
-		return nil
-	}
-	userName := struct {
-		Ip string `json:"ip"`
-	}{
-		Ip: ip,
-	}
-	userNameJson, err := json.Marshal(userName)
-	if err != nil {
-		log.Errorf("用户名序列化失败: %v", err)
-		return nil
-	}
-	base64UserName := base64.StdEncoding.EncodeToString(userNameJson)
-
-	mqttClientId := fmt.Sprintf("GID_test@@@%s@@@%s", deviceId, clientId)
-	pwd := util.Sha256Digest([]byte(mqttClientId))
-
-	return &MqttInfo{
-		Endpoint:       viper.GetString(otaConfigPrefix + "mqtt.endpoint"),
-		ClientId:       mqttClientId,
-		Username:       base64UserName,
-		Password:       pwd,
-		PublishTopic:   client.DeviceMockPubTopicPrefix,
-		SubscribeTopic: client.DeviceMockSubTopicPrefix,
 	}
 }
